test(ava): add eval datasets
