+++ src/index.ts
@@ -1,3 +1,2 @@
-/* advisor */
 export { Advisor } from './advisor';
 export { bindRenderer, type Spec } from './bind';
